Rename datahubhel.dhh_auth to datahubhel.auth
