+++ backend/datahubhel/auth/models.py
@@ -13,8 +13,8 @@ from django.db.models import QuerySet
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 
+from datahubhel.auth.utils import get_perm_obj
 from datahubhel.base_models import TimestampedUUIDModel
-from datahubhel.dhh_auth.utils import get_perm_obj
 
 
 class ClientPermission(TimestampedUUIDModel):
