--- backend/datahubhel/dhh_auth/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'datahubhel.dhh_auth.apps.DHHAuthConfig'
