+++ backend/datahubhel/gatekeeper/views.py
@@ -12,9 +12,9 @@ from django.urls import reverse
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from datahubhel.auth.models import ClientPermission
+from datahubhel.auth.utils import get_perm_obj
 from datahubhel.core.models import Datastream, Thing
-from datahubhel.dhh_auth.models import ClientPermission
-from datahubhel.dhh_auth.utils import get_perm_obj
 
 from .utils import (
     ENTITY_TO_DATASTREAM_PATH,
