+++ backend/datahubhel/urls.py
@@ -3,7 +3,7 @@ from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
 
-import datahubhel.dhh_auth.urls
+import datahubhel.auth.urls
 import datahubhel.gatekeeper.urls
 import datahubhel.mqttauth.urls
 import datahubhel.service.urls
@@ -12,7 +12,7 @@ import ta120_adapter.urls
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include(datahubhel.dhh_auth.urls)),
+    path('api/', include(datahubhel.auth.urls)),
     path('api/', include(datahubhel.service.urls)),
     path('api/', include(datahubhel.sta.urls)),
     path('api/', include(datahubhel.gatekeeper.urls)),
